# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Agent Service API
  version: 0.1.0
  description: Versioned public HTTP contract for the workflow and Agent platform.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /auth/access/members:
    get:
      tags: [AccessManagement]
      operationId: listMembers
      summary: List tenant members
      parameters: [{ $ref: '#/components/parameters/Keyword' }]
      responses:
        '200': { description: Members, content: { application/json: { schema: { $ref: '#/components/schemas/TenantMemberListApiResponse' } } } }
    post:
      tags: [AccessManagement]
      operationId: addMember
      summary: Add a member to the current tenant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddTenantMemberRequest' }
      responses:
        '200': { description: Member added, content: { application/json: { schema: { $ref: '#/components/schemas/TenantMemberApiResponse' } } } }
  /auth/access/members/{userId}/roles:
    post:
      tags: [AccessManagement]
      operationId: updateMemberRoles
      summary: Replace a member's tenant roles
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMemberRolesRequest' }
      responses:
        '200': { description: Roles updated, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
  /auth/access/members/{userId}/enabled:
    post:
      tags: [AccessManagement]
      operationId: updateMemberEnabled
      summary: Enable or disable a tenant member
      parameters:
        - { $ref: '#/components/parameters/UserId' }
        - { name: enabled, in: query, required: true, schema: { type: boolean } }
      responses:
        '200': { description: Member status updated, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
  /auth/access/roles:
    get:
      tags: [AccessManagement]
      operationId: listRoles
      summary: List tenant roles
      responses:
        '200': { description: Roles, content: { application/json: { schema: { $ref: '#/components/schemas/TenantRoleListApiResponse' } } } }
    post:
      tags: [AccessManagement]
      operationId: saveRole
      summary: Create or update a tenant role
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveTenantRoleRequest' }
      responses:
        '200': { description: Role saved, content: { application/json: { schema: { $ref: '#/components/schemas/TenantRoleApiResponse' } } } }
...
  /agents:
    get:
      tags: [Agent]
      operationId: listAgents
      summary: List Agent definitions
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Agent definitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDefinitionPageApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Agent]
      operationId: createAgent
      summary: Create an Agent definition
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentDefinitionRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDefinitionApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /agent-runs/manual-tests:
    post:
      tags: [Agent]
      operationId: submitAgentManualRun
      summary: Submit a manual Agent run
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualRunRequest'
      responses:
        '202':
          description: Run accepted for asynchronous execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunSubmissionApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /agents/{id}:
    delete:
      tags: [Agent]
      operationId: deleteAgent
      summary: Delete an unused Agent definition
      description: Deletes only a draft-only Agent definition with no execution history. Published definitions are retained for workflow and audit compatibility.
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/NumericId' }]
      responses:
        '200':
          description: Agent deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoidApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    post:
      tags: [Agent]
      operationId: updateAgent
      summary: Update an Agent definition
      parameters: [{ $ref: '#/components/parameters/NumericId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentDefinitionRequest' }
      responses:
        '200': { description: Agent updated, content: { application/json: { schema: { $ref: '#/components/schemas/AgentDefinitionApiResponse' } } } }
  /agents/{definitionId}/versions:
    get:
      tags: [Agent]
      operationId: listAgentVersions
      summary: List Agent versions
      parameters: [{ $ref: '#/components/parameters/NumericDefinitionId' }]
      responses:
        '200': { description: Agent versions, content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersionListApiResponse' } } } }
    post:
      tags: [Agent]
      operationId: createAgentDraft
      summary: Create an Agent draft version
      parameters: [{ $ref: '#/components/parameters/NumericDefinitionId' }]
      responses:
        '200': { description: Draft created, content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersionApiResponse' } } } }
  /agents/published-versions:
    get:
      tags: [Agent]
      operationId: pagePublishedAgentVersions
      summary: Search published Agent versions for workflow binding
      parameters:
        - { $ref: '#/components/parameters/PageNum' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Keyword' }
        - { name: versionId, in: query, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Published Agent versions, content: { application/json: { schema: { $ref: '#/components/schemas/PublishedAgentVersionPageApiResponse' } } } }
  /agents/{definitionId}/versions/{versionId}:
    post:
      tags: [Agent]
      operationId: updateAgentDraft
      summary: Update an Agent draft version
      parameters:
        - { $ref: '#/components/parameters/NumericDefinitionId' }
        - { $ref: '#/components/parameters/NumericVersionId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentVersionUpdateCommand' }
      responses:
        '200': { description: Draft updated, content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersionApiResponse' } } } }
  /agents/{definitionId}/versions/{versionId}/publish:
    post:
      tags: [Agent]
      operationId: publishAgentVersion
      summary: Publish an Agent version
      parameters:
        - { $ref: '#/components/parameters/NumericDefinitionId' }
        - { $ref: '#/components/parameters/NumericVersionId' }
      responses:
        '200': { description: Version published, content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersionApiResponse' } } } }
...
  /agent-runs:
    get:
      tags: [Agent]
      operationId: listAgentRuns
      summary: Page Agent runs
      parameters:
        - { $ref: '#/components/parameters/PageNum' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Keyword' }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200': { description: Agent runs, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunPageApiResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /agent-runs/{runId}:
    get:
      tags: [Agent]
      operationId: getAgentRun
      summary: Read an Agent run detail
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200': { description: Agent run detail, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunDetailApiResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /agent-runs/process-instances/{processInstanceId}:
    get:
      tags: [Agent]
      operationId: listAgentRunsByProcessInstance
      summary: List Agent runs for a workflow instance
      parameters:
        - { name: processInstanceId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Agent runs, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunListApiResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  parameters:
    PageNum:
      name: pageNum
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Keyword:
      name: keyword
      in: query
      schema: { type: string }
    NumericId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    NumericDefinitionId:
      name: definitionId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    NumericVersionId:
      name: versionId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: WORKFLOW_ACCESS_TOKEN
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        data: {}
        traceId: { type: string }
    VoidApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: object, nullable: true } }
...
    TenantMemberApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/TenantMember' } }
    TenantMemberListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: array, items: { $ref: '#/components/schemas/TenantMember' } } }
    TenantRoleApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/TenantRole' } }
    TenantRoleListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: array, items: { $ref: '#/components/schemas/TenantRole' } } }
...
    AgentDefinitionApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AgentDefinition' } }
    AgentDefinitionPageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AgentDefinitionPage' } }
    AgentVersionApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AgentVersion' } }
    AgentVersionListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: array, items: { $ref: '#/components/schemas/AgentVersion' } } }
    PublishedAgentVersionPageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/PublishedAgentVersionPage' } }
...
    AgentRunSubmissionApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AgentRunSubmission' } }
    AgentRunPageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AgentRunPage' } }
    AgentRunListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: array, items: { $ref: '#/components/schemas/AgentRun' } } }
    AgentRunDetailApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AgentRunDetail' } }
...
    TenantMember:
      type: object
      required: [userId, username, enabled, roles, globalRoles]
      properties:
        userId: { type: string }
        username: { type: string }
        displayName: { type: string, nullable: true }
        enabled: { type: boolean }
        roles: { type: array, uniqueItems: true, items: { type: string } }
        globalRoles: { type: array, uniqueItems: true, items: { type: string } }
        joinedAt: { type: string, format: date-time, nullable: true }
    TenantRole:
      type: object
      required: [roleCode, roleName, enabled, builtIn, permissions]
      properties:
        roleCode: { type: string }
        roleName: { type: string }
        description: { type: string, nullable: true }
        enabled: { type: boolean }
        builtIn: { type: boolean }
        permissions: { type: array, uniqueItems: true, items: { type: string } }
...
    PageMetadata:
      type: object
      required: [total, pageNum, pageSize, records]
      properties:
        total: { type: integer, format: int64 }
        pageNum: { type: integer }
        pageSize: { type: integer }
        records: { type: array, items: {} }
...
    AgentDefinition:
      type: object
      required: [id, code, name, enabled, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        enabled: { type: boolean }
        latestVersion: { type: integer, nullable: true }
        publishedVersion: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AgentDefinitionPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties: { records: { type: array, items: { $ref: '#/components/schemas/AgentDefinition' } } }
    AgentVersion:
      type: object
      required: [id, definitionId, version, status, executionMode, timeoutSeconds, failurePolicy]
      properties:
        id: { type: integer, format: int64 }
        definitionId: { type: integer, format: int64 }
        version: { type: integer }
        status: { type: string }
        executionMode: { type: string }
        providerId: { type: integer, format: int64, nullable: true }
        providerName: { type: string, nullable: true }
        modelName: { type: string, nullable: true }
        systemPrompt: { type: string, nullable: true }
        timeoutSeconds: { type: integer }
        failurePolicy: { type: string }
        inputSchema: { type: string, nullable: true }
        outputSchema: { type: string, nullable: true }
        createdBy: { type: string, nullable: true }
        publishedBy: { type: string, nullable: true }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        toolCodes: { type: array, items: { type: string, maxLength: 128 } }
    PublishedAgentVersion:
      type: object
      required: [id, definitionId, agentCode, agentName, version, executionMode, timeoutSeconds, contractFingerprint]
      properties:
        id: { type: integer, format: int64 }
        definitionId: { type: integer, format: int64 }
        agentCode: { type: string }
        agentName: { type: string }
        version: { type: integer }
        executionMode: { type: string }
        timeoutSeconds: { type: integer }
        inputSchema: { type: string, nullable: true }
        outputSchema: { type: string, nullable: true }
        contractFingerprint: { type: string }
    PublishedAgentVersionPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties: { records: { type: array, items: { $ref: '#/components/schemas/PublishedAgentVersion' } } }
...
    AgentRunSubmission:
      type: object
      required: [runId, status]
      properties:
        runId: { type: integer, format: int64 }
        status: { type: string }
    AgentRun:
      type: object
      required: [id, agentCode, agentName, agentVersion, status, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        agentCode: { type: string }
        agentName: { type: string }
        agentVersion: { type: integer }
        status: { type: string }
        resultStatus: { type: string, nullable: true }
        processInstanceId: { type: string, nullable: true }
        activityId: { type: string, nullable: true }
        errorCode: { type: string, nullable: true }
        deadlineAt: { type: string, format: date-time, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AgentRunPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties: { records: { type: array, items: { $ref: '#/components/schemas/AgentRun' } } }
    AgentRunDetail:
      type: object
      required: [run, payload, attempts, steps, modelInvocations, checkpoints, stateHistory]
      properties:
        run: { $ref: '#/components/schemas/AgentRun' }
        payload: { type: object }
        attempts: { type: array, items: { type: object } }
        steps: { type: array, items: { type: object } }
        modelInvocations: { type: array, items: { type: object } }
        checkpoints: { type: array, items: { type: object } }
        stateHistory: { type: array, items: { type: object } }
...
    AddTenantMemberRequest:
      type: object
      required: [userId]
      properties:
        userId: { type: string }
        roleCodes: { type: array, items: { type: string } }
    UpdateMemberRolesRequest:
      type: object
      required: [roleCodes]
      properties:
        roleCodes: { type: array, items: { type: string } }
    SaveTenantRoleRequest:
      type: object
      required: [roleCode, roleName, permissionCodes]
      properties:
        roleCode: { type: string }
        roleName: { type: string }
        description: { type: string }
        permissionCodes: { type: array, items: { type: string } }
...
    AgentDefinitionRequest:
      type: object
      required: [agentCode, agentName]
      properties:
        agentCode: { type: string }
        agentName: { type: string }
        description: { type: string }
...
    AgentVersionUpdateCommand:
      type: object
      required: [timeoutSeconds, failurePolicy]
      properties:
        executionMode:
          type: string
          enum: [MODEL_ONLY, PLATFORM_AGENT, REMOTE_AGENT]
          default: MODEL_ONLY
        providerId: { type: integer, format: int64 }
        modelName: { type: string, maxLength: 128 }
        systemPrompt: { type: string, maxLength: 20000 }
        timeoutSeconds: { type: integer, minimum: 1, maximum: 3600 }
        failurePolicy:
          type: string
          enum: [FAIL_PROCESS, CONTINUE_EMPTY, MANUAL_REVIEW]
        inputSchema: { type: string, maxLength: 20000 }
        outputSchema: { type: string, maxLength: 20000 }
        toolCodes: { type: array, items: { type: string, maxLength: 128 }, maxItems: 128 }
...
    ManualRunRequest:
      type: object
      required: [agentVersionId, input]
      properties:
        agentVersionId: { type: integer, format: int64 }
        input: { type: string }
        idempotencyKey: { type: string }
  responses:
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    NotFound:
      description: Resource does not exist in the current tenant scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    Conflict:
      description: Operation conflicts with the current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }